refactored get_remote_components()
